requirements.txt
dj-database-url==0.4.1
dj-static==0.0.6
Django==1.10.2
django-cors-headers==1.2.2
djangorestframework==3.5.1
gunicorn==19.6.0
Markdown==2.6.7
psycopg2==2.6.2
PyYAML==3.12
static3==0.7.0
